Broke out functions into module
